Convert to umbrella project with `ex_ttrpg_dev` and `ttrpg_dev_cli` apps
Splits the CLI into its own package (`ttrpg_dev_cli`) so it can be
versioned and distributed independently from the core library. The CLI
app depends on the core via `in_umbrella: true`. Build the escript with
`mix escript` (alias for `mix do --app ttrpg_dev_cli escript.build`).
