+++ .formatter.exs
@@ -1,4 +1,3 @@
-# Used by "mix format"
 [
-  inputs: ["{mix,.formatter}.exs", "{config,lib,test}/**/*.{ex,exs}"]
+  subdirectories: ["apps/*"]
 ]
