+++ apps/ex_ttrpg_dev/lib/characters/character.ex
