+++ apps/ex_ttrpg_dev/lib/characters/metadata.ex
