+++ apps/ex_ttrpg_dev/lib/globals.ex
