+++ apps/ex_ttrpg_dev/lib/rule_system/evaluator.ex
