+++ apps/ex_ttrpg_dev/lib/rule_system/graph.ex
