+++ apps/ex_ttrpg_dev/lib/rule_system/rule_module.ex
