+++ apps/ex_ttrpg_dev/priv/system_configs/dnd_5e_srd/attributes.toml
