+++ apps/ex_ttrpg_dev/test/globals_test.exs
