+++ apps/ex_ttrpg_dev/test/rule_system/evaluator_test.exs
