+++ apps/ttrpg_dev_cli/lib/cli/character_display.ex
