+++ apps/ttrpg_dev_cli/lib/cli/generate.ex
