+++ apps/ttrpg_dev_cli/test/cli/character_display_test.exs
