+++ mix.exs
@@ -1,54 +1,22 @@
-defmodule ExTTRPGDev.MixProject do
+defmodule ExTtrpgDevUmbrella.MixProject do
   use Mix.Project
 
   def project do
     [
-      app: :ex_ttrpg_dev,
+      apps_path: "apps",
       version: "0.6.0",
       elixir: "~> 1.17",
       start_permanent: Mix.env() == :prod,
       deps: deps(),
-      escript: escript(),
-      description: description(),
-      package: package(),
-      name: "ExTTRPGDev",
-      source_url: "https://github.com/TTRPG-Dev/ex_ttrpg_dev"
+      aliases: aliases()
     ]
   end
 
-  def escript do
-    [main_module: ExTTRPGDev.CLI, include_priv: [:ex_ttrpg_dev]]
-  end
-
-  # Run "mix help compile.app" to learn about applications.
-  def application do
+  defp aliases do
     [
-      extra_applications: [:logger]
+      escript: ["do --app ttrpg_dev_cli escript.build"]
     ]
   end
 
-  # Run "mix help deps" to learn about dependencies.
-  defp deps do
-    [
-      {:optimus, "~> 0.6"},
-      {:ex_doc, "~> 0.40", only: :dev, runtime: false},
-      {:poison, "~> 6.0"},
-      {:credo, "~> 1.7", only: [:dev, :test], runtime: false},
-      {:faker, "~> 0.18"},
-      {:toml_elixir, "~> 3.1"},
-      {:abacus, "~> 2.1"},
-      {:libgraph, "~> 0.16"}
-    ]
-  end
-
-  defp description() do
-    "ExTTRPGDev is a general utility for tabletop role-playing games."
-  end
-
-  defp package() do
-    [
-      licenses: ["GPL-3.0-only"],
-      links: %{"GitHub" => "https://github.com/TTRPG-Dev/ex_ttrpg_dev"}
-    ]
-  end
+  defp deps, do: []
 end
